# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "APIResponse_AnnotationResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnnotationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload."
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[AnnotationResponse]",
        "type": "object"
      },
...
      "APIResponse_NoneType_": {
        "properties": {
          "data": {
            "description": "Endpoint specific payload.",
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[NoneType]",
        "type": "object"
      },
...
      "APIResponse_list_AnnotationResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[AnnotationResponse]]",
        "type": "object"
      },
      "APIResponse_list_AnnotationReviewResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AnnotationReviewResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Endpoint specific payload.",
            "title": "Data"
          },
          "message": {
            "default": "Request completed successfully.",
            "description": "Human readable outcome of the request.",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "True when the request completed successfully.",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "APIResponse[list[AnnotationReviewResponse]]",
        "type": "object"
      },
...
      "AnnotationCreate": {
        "description": "Payload used by an annotator to submit a label.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Annotator confidence between 0 and 1.",
            "title": "Confidence"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "task_id": {
            "exclusiveMinimum": 0.0,
            "title": "Task Id",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "label"
        ],
        "title": "AnnotationCreate",
        "type": "object"
      },
      "AnnotationResponse": {
        "description": "Annotation representation returned by the API.",
        "properties": {
          "ai_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Confidence"
          },
          "ai_suggested_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Label"
          },
          "annotator_id": {
            "title": "Annotator Id",
            "type": "integer"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "annotator_id",
          "label",
          "notes",
          "confidence",
          "ai_suggested_label",
          "ai_confidence",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AnnotationResponse",
        "type": "object"
      },
...
      "AnnotationReviewResponse": {
        "description": "Quality review representation returned by the API.",
        "properties": {
          "annotation_id": {
            "title": "Annotation Id",
            "type": "integer"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reviewer_id": {
            "title": "Reviewer Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "annotation_id",
          "reviewer_id",
          "decision",
          "comment",
          "created_at"
        ],
        "title": "AnnotationReviewResponse",
        "type": "object"
      },
...
      "AnnotationUpdate": {
        "description": "Payload used to revise an existing annotation.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "AnnotationUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the AI Dataset Labeling Marketplace: datasets, labeling jobs, annotation tasks, annotations, quality review, progress tracking, dataset export and the AI annotation assistant.",
    "title": "AI Dataset Labeling Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/annotation-reviews/annotation/{annotation_id}": {
      "get": {
        "description": "Return every decision recorded for one annotation.",
        "operationId": "get_reviews_for_annotation_api_annotation_reviews_annotation__annotation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationReviewResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the review history of an annotation",
        "tags": [
          "Annotation Reviews"
        ]
      }
    },
...
    "/api/annotations": {
      "get": {
        "description": "Return every annotation visible to the authenticated role.",
        "operationId": "get_annotations_api_annotations_get",
        "parameters": [
          {
            "in": "query",
            "name": "job_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List annotations",
        "tags": [
          "Annotations"
        ]
      },
      "post": {
        "description": "Submit an annotation, storing the AI suggestion for agreement metrics.",
        "operationId": "submit_annotation_api_annotations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AnnotationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit a label for an annotation task",
        "tags": [
          "Annotations"
        ]
      }
    },
    "/api/annotations/job/{job_id}": {
      "get": {
        "description": "Return every annotation submitted for a job.",
        "operationId": "get_annotations_for_job_api_annotations_job__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the annotations of a labeling job",
        "tags": [
          "Annotations"
        ]
      }
    },
    "/api/annotations/task/{task_id}": {
      "get": {
        "description": "Return every annotation attempt recorded for a task.",
        "operationId": "get_annotations_for_task_api_annotations_task__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_AnnotationResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the annotation history of a task",
        "tags": [
          "Annotations"
        ]
      }
    },
    "/api/annotations/{annotation_id}": {
      "delete": {
        "description": "Withdraw an annotation that has not been approved.",
        "operationId": "delete_annotation_api_annotations__annotation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw a submitted annotation",
        "tags": [
          "Annotations"
        ]
      },
      "get": {
        "description": "Return a single annotation by identifier.",
        "operationId": "get_annotation_api_annotations__annotation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AnnotationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve one annotation",
        "tags": [
          "Annotations"
        ]
      },
      "put": {
        "description": "Revise a label that has not been approved yet.",
        "operationId": "update_annotation_api_annotations__annotation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AnnotationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revise a submitted annotation",
        "tags": [
          "Annotations"
        ]
      }
    },
...
  }
}